+++ log/log_test.go
@@ -1,14 +1,12 @@
 package log
 
 import (
-	"bytes"
 	"errors"
 	"io"
 	"testing"
 	"time"
 
 	"github.com/sirupsen/logrus"
-	logrustest "github.com/sirupsen/logrus/hooks/test"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -120,88 +118,6 @@ func TestJSONFormatterErrorHandling(t *testing.T) {
 	})
 }
 
-func Test_Logger(t *testing.T) {
-	makeDummySink := func(writer io.Writer) Sinker {
-		return NewSink(writer, &logrus.TextFormatter{}, AcceptorAllowAll)
-	}
-
-	t.Run("Setup", func(t *testing.T) {
-		t.Run("flushes setup", func(t *testing.T) {
-			buf := &bytes.Buffer{}
-
-			lgr := New()
-			lgr.Info("pre-setup log")
-
-			assert.Len(t, lgr.tmpLogsCollector.entries, 1)
-			lgr.Setup(func(b *Builder) {
-				b.AddSink(makeDummySink(buf))
-			})
-
-			assert.Empty(t, lgr.tmpLogsCollector.entries)
-			assert.Contains(t, buf.String(), "pre-setup log")
-		})
-
-		t.Run("panics if called twice", func(t *testing.T) {
-			lgr := New()
-			lgr.Setup(func(_ *Builder) {})
-
-			assert.Panics(t, func() {
-				lgr.Setup(func(_ *Builder) {})
-			})
-		})
-	})
-
-	t.Run("flushes to emergency logger if setup was not called", func(t *testing.T) {
-		buf := bytes.Buffer{}
-
-		lgr := New()
-		lgr.emergencyLogger.SetOutput(&buf)
-
-		lgr.Info("fatal startup error")
-		assert.Len(t, lgr.tmpLogsCollector.entries, 1)
-
-		lgr.Flush()
-
-		assert.Empty(t, lgr.tmpLogsCollector.entries)
-		assert.Contains(t, buf.String(), "fatal startup error")
-	})
-
-	t.Run("Flush", func(t *testing.T) {
-		t.Run("does not add logs to output", func(t *testing.T) {
-			lgr := New()
-			emBuf := &bytes.Buffer{}
-			lgr.emergencyLogger.SetOutput(emBuf)
-
-			lgr.Setup(func(_ *Builder) {})
-
-			log.Info("post-setup log")
-			lgr.Flush()
-
-			assert.Empty(t, emBuf.String())
-		})
-	})
-
-	t.Run("RemoveHook", func(t *testing.T) {
-		t.Run("removes hook", func(t *testing.T) {
-			logger := New()
-			logger.ReplaceHooks(make(logrus.LevelHooks))
-
-			hook := &logrustest.Hook{}
-			logger.AddHook(hook)
-
-			for _, hooks := range logger.Hooks {
-				assert.True(t, len(hooks) == 1, "each logger level hes it's hook")
-			}
-
-			logger.RemoveHook(hook)
-
-			for _, hooks := range logger.Hooks {
-				assert.True(t, len(hooks) == 0, "has removed all the hooks from logger")
-			}
-		})
-	})
-}
-
 func TestGetAndGetRaw(t *testing.T) {
 	logger := Get()
 	assert.NotNil(t, logger)
@@ -225,10 +141,10 @@ func Test_Logger_IsLegacyFormatter(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			lgr := New()
 			lgr.Setup(func(b *Builder) {
-				b.SetLegacyLogformat(tc.format == FormatLegacy)
+				b.SetLogformat(tc.format)
 			})
 
-			assert.Equal(t, tc.expectedResult, lgr.IsLegacyFormatter())
+			assert.Equal(t, tc.expectedResult, lgr.IsLegacyFormatterEnabled())
 		})
 	}
 }
